Moved Sensor Upgrade Checks into program_start_checks.py instead of using a linux service  Moved program_upgrade_functions.py into program_start_checks.py  Auto restart service only if running from the service folder Allow running upgrade checks without root Set variable for new install version
