+++ extras/deb_package_files/DEBIAN/control
@@ -1,5 +1,5 @@
 Package: kootnet-sensors
-Version: 0.29-84
+Version: 0.29-86
 Section: utils
 Priority: extra
 Architecture: all
