--- extras/deb_package_files/etc/systemd/system/SensorUpgradeChecks.service
@@ -1,12 +1,0 @@
-[Unit]
-Description=KootNet Sensors - Upgrade Checks
-
-[Service]
-Type=simple
-Restart=no
-User=root
-Group=root
-ExecStart=/home/kootnet_data/env/bin/python /opt/kootnet-sensors/run_upgrade_checks.py
-
-[Install]
-WantedBy=multi-user.target
