+++ operations_modules/program_start_checks.py
@@ -22,7 +22,11 @@ from operations_modules import logger
 from operations_modules import file_locations
 from operations_modules import software_version
 from operations_modules import os_cli_commands
+from operations_modules.app_generic_functions import thread_function
 from operations_modules.sqlite_database import check_database_structure
+from configuration_modules.config_primary import CreatePrimaryConfiguration
+from configuration_modules.config_installed_sensors import CreateInstalledSensorsConfiguration
+from configuration_modules.config_trigger_variances import CreateTriggerVariancesConfiguration
 
 
 def run_program_start_checks():
@@ -34,14 +38,13 @@ def run_program_start_checks():
     _set_file_permissions()
     check_database_structure()
     _check_ssl_files()
-    if os.geteuid() == 0:
-        if software_version.old_version != software_version.version:
-            logger.primary_logger.info(" -- Starting Programs Upgrade Checks ...")
-            os.system("systemctl start SensorUpgradeChecks")
-            # Sleep before loading anything due to needed updates
-            # The update service will automatically restart this app when it's done
-            while True:
-                time.sleep(30)
+    if software_version.old_version != software_version.version:
+        logger.primary_logger.info(" -- Starting Upgrade Checks ...")
+        thread_function(_run_upgrade_checks)
+        # Sleep before loading anything due to needed updates
+        # The update function will automatically restart the service
+        while True:
+            time.sleep(30)
 
 
 def _set_file_permissions():
@@ -78,3 +81,68 @@ def _check_ssl_files():
         terminal_command_part1 = "openssl x509 -req -days 3650 -in " + file_locations.http_ssl_csr
         terminal_command_part2 = " -signkey " + file_locations.http_ssl_key + " -out " + file_locations.http_ssl_crt
         os.system(terminal_command_part1 + terminal_command_part2)
+
+
+def _run_upgrade_checks():
+    """
+     Checks previous written version of the program to the current version.
+     If the current version is different, start upgrade functions.
+    """
+    msg = "Old Version: " + software_version.old_version + " || New Version: " + software_version.version
+    logger.primary_logger.debug(msg)
+    previous_version = software_version.CreateRefinedVersion(software_version.old_version)
+    no_changes = True
+
+    if previous_version.major_version == "New_Install":
+        logger.primary_logger.info("New Install Detected")
+        no_changes = False
+    elif previous_version.major_version == "Beta":
+        if previous_version.feature_version == 29:
+            if previous_version.minor_version < 13:
+                reset_installed_sensors()
+                reset_variance_config()
+    elif previous_version.major_version == "Alpha":
+        msg = "Upgraded: " + software_version.old_version + " || New: " + software_version.version
+        logger.primary_logger.info(msg)
+        no_changes = False
+        reset_installed_sensors()
+        reset_primary_config()
+        reset_variance_config()
+    else:
+        no_changes = False
+        msg = "Bad or Missing Previous Version Detected - Resetting Config and Installed Sensors"
+        logger.primary_logger.error(msg)
+        reset_installed_sensors()
+        reset_primary_config()
+
+    # Since run_upgrade_checks is only run if there is a different version, show upgrade but no configuration changes
+    if no_changes:
+        msg = "Upgrade detected || No configuration changes || Old: "
+        logger.primary_logger.info(msg + software_version.old_version + " New: " + software_version.version)
+    software_version.write_program_version_to_file()
+    if file_locations.program_root_dir == "/opt/kootnet-sensors":
+        logger.primary_logger.info("Upgrade Complete - Restarting the Sensor service")
+        os.system(os_cli_commands.restart_sensor_services_command)
+    else:
+        logger.primary_logger.info("Upgrade Complete - Please restart the Sensor program")
+
+
+def reset_installed_sensors():
+    """ Writes a default installed sensor configuration file. """
+    logger.primary_logger.warning(" **** Installed Sensors Configuration Reset ****")
+    default_installed_sensors = CreateInstalledSensorsConfiguration(load_from_file=False)
+    default_installed_sensors.save_config_to_file()
+
+
+def reset_primary_config():
+    """ Writes a default main configuration file. """
+    logger.primary_logger.warning(" **** Main Configuration Reset ****")
+    default_primary_config = CreatePrimaryConfiguration(load_from_file=False)
+    default_primary_config.save_config_to_file()
+
+
+def reset_variance_config():
+    """ Writes a default Trigger Variance configuration file. """
+    logger.primary_logger.warning(" **** Trigger Variances Configuration Reset ****")
+    default_variance = CreateTriggerVariancesConfiguration(load_from_file=False)
+    default_variance.save_config_to_file()
